KL-123/refactor : change directory
subcategory 디렉토리에서 category 디렉토리로 subcategory 관련 파일들을 이동.
