+++ src/main/java/taco/klkl/domain/category/convert/SubcategoryNameConverter.java
@@ -1,8 +1,8 @@
-package taco.klkl.domain.subcategory.convert;
+package taco.klkl.domain.category.convert;
 
 import jakarta.persistence.AttributeConverter;
 import jakarta.persistence.Converter;
-import taco.klkl.domain.subcategory.domain.SubcategoryName;
+import taco.klkl.domain.category.domain.SubcategoryName;
 
 @Converter(autoApply = true)
 public class SubcategoryNameConverter implements AttributeConverter<SubcategoryName, String> {
