+++ src/main/java/taco/klkl/domain/category/dao/SubcategoryRepository.java
@@ -1,13 +1,10 @@
-package taco.klkl.domain.subcategory.dao;
-
-import java.util.List;
+package taco.klkl.domain.category.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import taco.klkl.domain.subcategory.domain.Subcategory;
+import taco.klkl.domain.category.domain.Subcategory;
 
 @Repository
 public interface SubcategoryRepository extends JpaRepository<Subcategory, Long> {
-	List<Subcategory> findAllByCategoryId(Long id);
 }
