+++ src/main/java/taco/klkl/domain/category/domain/Subcategory.java
@@ -1,7 +1,8 @@
-package taco.klkl.domain.subcategory.domain;
+package taco.klkl.domain.category.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
+import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
@@ -10,7 +11,6 @@ import jakarta.persistence.ManyToOne;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import taco.klkl.domain.category.domain.Category;
 
 @Getter
 @Entity
@@ -21,7 +21,7 @@ public class Subcategory {
 	@GeneratedValue(strategy = GenerationType.AUTO)
 	private Long id;
 
-	@ManyToOne
+	@ManyToOne(fetch = FetchType.LAZY)
 	@JoinColumn(name = "category_id")
 	private Category category;
